Add first byte as serialization type to WKB.
